+++ src/main/kotlin/data/registers/GeneralRegisters.kt
@@ -11,13 +11,13 @@ import kotlin.system.exitProcess
  *
  * This class provides methods for reading and writing values to the individual general-purpose registers (G1-G4).
  *
- * General-purpose registers are used for storing operands, intermediate results, and other values during program execution.
+ * General-purpose registers are used for storing operands, Longermediate results, and other values during program execution.
  */
 class GeneralRegisters {
-    var g1: Int? = null
-    var g2: Int? = null
-    var g3: Int? = null
-    var g4: Int? = null
+    var g1: Long? = null
+    var g2: Long? = null
+    var g3: Long? = null
+    var g4: Long? = null
 
     /**
      * Reads the value from the specified general purpose register.
@@ -25,7 +25,7 @@ class GeneralRegisters {
      * @param registers The general purpose register to read from.
      * @return The value stored in the specified register.
      */
-    fun read(registers: GeneralRegisterType): Int {
+    fun read(registers: GeneralRegisterType): Long {
         try {
             return when (registers) {
                 GeneralRegisterType.G1 -> g1!!
@@ -45,7 +45,7 @@ class GeneralRegisters {
      * @param registers The general purpose register to write to.
      * @param value The value to write to the register.
      */
-    fun write(registers: GeneralRegisterType, value: Int) {
+    fun write(registers: GeneralRegisterType, value: Long) {
         when (registers) {
             GeneralRegisterType.G1 -> g1 = value
             GeneralRegisterType.G2 -> g2 = value
