+++ src/main/kotlin/helpers/stingToMemoryAddress.kt
@@ -3,5 +3,5 @@ package org.example.helpers
 import org.example.data.memory.MemoryAddress
 
 fun String.toMemoryAddress(): MemoryAddress {
-    return MemoryAddress(this.toInt())
+    return MemoryAddress(this.toLong())
 }
