+++ src/main/kotlin/kvmInternals/systemCalls/calls/exit.kt
@@ -7,6 +7,6 @@ import kotlin.system.exitProcess
 
 fun SystemCall.exit(s2: SuperRegisterType) {
     val exitCode = fullRegisterRead(s2)
-    exitProcess(exitCode)
+    exitProcess(exitCode.toInt())
 
 }
