+++ src/main/kotlin/kvmInternals/systemCalls/calls/getpid.kt
@@ -5,6 +5,6 @@ import org.example.helpers.fullRegisterWrite
 import org.example.kvmInternals.systemCalls.SystemCall
 
 fun SystemCall.getPid() {
-    fullRegisterWrite(SuperRegisterType.R2, ProcessHandle.current().pid().toInt())
+    fullRegisterWrite(SuperRegisterType.R2, ProcessHandle.current().pid())
 
 }
